# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Next API Backend
  version: 1.0.0
  description: Rebuilt backend APIs (MongoDB Atlas + Next.js route handlers)
servers:
  - url: /
security: []
...
paths:
  /api/contact:
    post:
      tags: [Contact]
      summary: Send contact notification and user confirmation email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
      responses:
        '200':
          description: Email workflow succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactSuccess'
        '400':
          $ref: '#/components/responses/Error400'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
        '502':
          $ref: '#/components/responses/Error502'
  /api/revalidate:
    get:
      tags: [Revalidate]
      summary: Revalidate an ISR path
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: secret
          required: true
          schema: { type: string }
        - in: query
          name: path
          required: false
          schema: { type: string, default: / }
        - in: query
          name: type
          required: false
          schema: { type: string, enum: [page, layout], default: page }
      responses:
        '200':
          description: Revalidation accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevalidateSuccess'
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
...
  /api/tools/image-to-pdf/share/{token}:
    get:
      tags: [Tools]
      summary: Resolve metadata for a temporary share token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Share metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageToPdfShareMetadataSuccess'
        '404':
          $ref: '#/components/responses/Error404'
        '410':
          $ref: '#/components/responses/Error410'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags: [Tools]
      summary: Revoke and delete a temporary share token (owner fingerprint/IP)
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Token revoked and asset deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteSuccess' }
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/tools/image-to-pdf/share/{token}/download:
    get:
      tags: [Tools]
      summary: Download PDF for a temporary share token
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      responses:
        '200':
          description: PDF stream
          headers:
            Content-Type: { schema: { type: string } }
            Content-Disposition: { schema: { type: string } }
            Cache-Control: { schema: { type: string } }
          content:
            application/pdf:
              schema: { type: string, format: binary }
        '404':
          $ref: '#/components/responses/Error404'
        '410':
          $ref: '#/components/responses/Error410'
        '429':
          $ref: '#/components/responses/Error429'
        '500':
          $ref: '#/components/responses/Error500'
  /api/tools/image-to-pdf/cleanup:
    post:
      tags: [Tools]
      summary: Cleanup expired temporary PDF assets (cron/internal)
      parameters:
        - in: header
          name: x-tools-cleanup-secret
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cleanup sweep completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageToPdfCleanupSuccess'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error404:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error403:
      description: Forbidden
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error410:
      description: Gone (expired/revoked)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Error429:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error500:
      description: Internal error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error502:
      description: Upstream provider error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
    ContactRequest:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string }
        email: { type: string, format: email }
        subject: { type: string }
        message: { type: string }
    ContactSuccess:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, enum: [true] }
        message: { type: string }
        id: { type: string, nullable: true }
    RevalidateSuccess:
      type: object
      required: [success, revalidated, path, type, timestamp]
      properties:
        success: { type: boolean, enum: [true] }
        revalidated: { type: boolean, enum: [true] }
        path: { type: string }
        type: { type: string, enum: [page, layout] }
        timestamp: { type: string, format: date-time }
...
    ImageToPdfShareMetadataSuccess:
      type: object
      required:
        - success
        - assetId
        - filename
        - pageCount
        - fileSizeBytes
        - createdAt
        - expiresAt
      properties:
        success: { type: boolean, enum: [true] }
        assetId: { type: string }
        filename: { type: string }
        pageCount: { type: integer, minimum: 1 }
        fileSizeBytes: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
    ImageToPdfCleanupSuccess:
      type: object
      required: [success, now, removedAssets, removedChunks, removedTokens]
      properties:
        success: { type: boolean, enum: [true] }
        now: { type: string, format: date-time }
        removedAssets: { type: integer, minimum: 0 }
        removedChunks: { type: integer, minimum: 0 }
        removedTokens: { type: integer, minimum: 0 }